+++ pipelex/core/bundles/exceptions.py
@@ -3,60 +3,6 @@ from pydantic import BaseModel, Field
 from pipelex.types import StrEnum
 
 
-class PipeValidationErrorType(StrEnum):
-    """Types of pipe validation errors.
-
-    These error types are raised during pipe validation from Pipe/Concept classes.
-    Only some are auto-fixed in the builder loop (marked below).
-    """
-
-    # Errors that are auto-fixed in builder_loop.py
-    MISSING_INPUT_VARIABLE = "missing_input_variable"  # AUTO-FIXED
-    EXTRANEOUS_INPUT_VARIABLE = "extraneous_input_variable"  # AUTO-FIXED
-    INPUT_REQUIREMENT_MISMATCH = "input_requirement_mismatch"  # AUTO-FIXED
-    INADEQUATE_OUTPUT_CONCEPT = "inadequate_output_concept"  # AUTO-FIXED
-
-    # Errors that are raised but NOT auto-fixed (will fail validation)
-    LLM_OUTPUT_CANNOT_BE_IMAGE = "llm_output_cannot_be_image"
-    IMG_GEN_INPUT_NOT_TEXT_COMPATIBLE = "img_gen_input_not_text_compatible"
-
-    # Generic fallback for unexpected validation errors
-    UNKNOWN_VALIDATION_ERROR = "unknown_validation_error"
-
-
-class PipesAndConceptValidationErrorData(BaseModel):
-    """Structured validation error data for Pipe/Concept validation errors.
-
-    This model captures validation errors raised by Pipe or Concept classes during
-    their validation (NOT blueprint validation errors).
-
-    These errors come from:
-    - PipeAbstract and its subclasses (PipeLLM, PipeExtract, etc.)
-    - Concept validation
-    """
-
-    # === Source Context ===
-    domain: str | None = Field(None, description="Domain where error occurred")
-    source: str | None = Field(None, description="Source file path")
-
-    # === Entity Context (what failed) ===
-    pipe_code: str | None = Field(None, description="Pipe code if error is in a pipe")
-    concept_code: str | None = Field(None, description="Concept code if error is in a concept")
-    field_name: str | None = Field(None, description="Specific field that failed")
-
-    # === Error Classification ===
-    error_type: PipeValidationErrorType = Field(
-        description="Type of pipe/concept validation error",
-    )
-
-    # === Error Details ===
-    message: str = Field(description="Human-readable error message")
-    field_path: str = Field(description="Path to field in dot notation")
-
-    # === Variable names for input/output errors ===
-    variable_names: list[str] | None = Field(None, description="Variable names (for input errors)")
-
-
 class PipelexBundleBlueprintFixableErrorType(StrEnum):
     """Types of fixable validation errors in Pipelex bundle blueprints.
