+++ pipelex/core/pipes/inputs/exceptions.py
@@ -14,3 +14,7 @@ class PipeInputError(PipelexError):
         self.variable_name = variable_name
         self.concept_code = concept_code
         super().__init__(message)
+
+
+class PipeInputNotFoundError(PipelexError):
+    pass
