+++ pipelex/core/pipes/inputs/input_requirements.py
@@ -4,7 +4,7 @@ from pydantic import BaseModel, Field, RootModel, field_validator
 
 from pipelex import log
 from pipelex.core.concepts.concept import Concept
-from pipelex.core.pipes.exceptions import PipeInputNotFoundError
+from pipelex.core.pipes.inputs.exceptions import PipeInputNotFoundError
 from pipelex.core.pipes.variable_multiplicity import VariableMultiplicity
 from pipelex.core.stuffs.stuff_content import StuffContent
