+++ pipelex/pipe_controllers/batch/pipe_batch.py
@@ -7,8 +7,7 @@ from typing_extensions import override
 from pipelex.config import get_config
 from pipelex.core.memory.exceptions import WorkingMemoryStuffNotFoundError
 from pipelex.core.memory.working_memory import MAIN_STUFF_NAME, WorkingMemory
-from pipelex.core.pipes.exceptions import PipeInputNotFoundError
-from pipelex.core.pipes.inputs.exceptions import PipeInputError
+from pipelex.core.pipes.inputs.exceptions import PipeInputError, PipeInputNotFoundError
 from pipelex.core.pipes.inputs.input_requirements import InputRequirements
 from pipelex.core.pipes.pipe_output import PipeOutput
 from pipelex.core.stuffs.list_content import ListContent
