+++ pipelex/pipe_operators/extract/pipe_extract.py
@@ -11,13 +11,13 @@ from pipelex.cogt.extract.extract_input import ExtractInput
 from pipelex.cogt.extract.extract_job_components import ExtractJobConfig, ExtractJobParams
 from pipelex.cogt.extract.extract_setting import ExtractModelChoice, ExtractSetting
 from pipelex.cogt.models.model_deck_check import check_extract_choice_with_deck
-from pipelex.core.bundles.exceptions import PipeValidationErrorType
 from pipelex.core.concepts.native.concept_native import NativeConceptCode
 from pipelex.core.memory.working_memory import WorkingMemory
 from pipelex.core.pipe_errors import PipeDefinitionError
 from pipelex.core.pipes.exceptions import PipeValidationError
 from pipelex.core.pipes.inputs.input_requirements import InputRequirements
 from pipelex.core.pipes.pipe_output import PipeOutput
+from pipelex.core.pipes.validation import PipeValidationErrorType
 from pipelex.core.stuffs.image_content import ImageContent
 from pipelex.core.stuffs.list_content import ListContent
 from pipelex.core.stuffs.page_content import PageContent
@@ -86,13 +86,14 @@ class PipeExtract(PipeOperator[PipeExtractOutput]):
     @override
     def validate_output_with_library(self):
         if self.output != get_native_concept(native_concept=NativeConceptCode.PAGE):
+            msg = f"PipeExtract output should be a Page concept, but is {self.output.concept_string}"
             raise PipeValidationError(
+                message=msg,
                 error_type=PipeValidationErrorType.INADEQUATE_OUTPUT_CONCEPT,
                 domain=self.domain,
                 pipe_code=self.code,
                 provided_concept_code=self.output.concept_string,
                 required_concept_codes=[NativeConceptCode.PAGE.concept_string],
-                explanation=f"PipeExtract output should be a Page concept, but is {self.output.concept_string}",
             )
 
     @override
