+++ pipelex/pipe_operators/img_gen/pipe_img_gen.py
@@ -11,7 +11,6 @@ from pipelex.cogt.img_gen.img_gen_prompt import ImgGenPrompt
 from pipelex.cogt.img_gen.img_gen_setting import ImgGenModelChoice, ImgGenSetting
 from pipelex.cogt.models.model_deck_check import check_img_gen_choice_with_deck
 from pipelex.config import get_config
-from pipelex.core.bundles.exceptions import PipeValidationErrorType
 from pipelex.core.concepts.concept_factory import ConceptFactory
 from pipelex.core.concepts.native.concept_native import NativeConceptCode
 from pipelex.core.memory.exceptions import WorkingMemoryStuffNotFoundError
@@ -20,6 +19,7 @@ from pipelex.core.pipes.exceptions import PipeValidationError
 from pipelex.core.pipes.inputs.input_requirements import InputRequirements
 from pipelex.core.pipes.inputs.input_requirements_factory import InputRequirementsFactory
 from pipelex.core.pipes.pipe_output import PipeOutput
+from pipelex.core.pipes.validation import PipeValidationErrorType
 from pipelex.core.pipes.variable_multiplicity import VariableMultiplicity
 from pipelex.core.stuffs.exceptions import StuffContentTypeError
 from pipelex.core.stuffs.image_content import ImageContent
@@ -124,18 +124,18 @@ class PipeImgGen(PipeOperator[PipeImgGenOutput]):
             wanted_concept=ConceptFactory.make_native_concept(native_concept_code=NativeConceptCode.TEXT),
             strict=True,
         ):
-            explanation = (
+            msg = (
                 f"The input of a PipeImgGen must be compatible with the Text concept (or refine it). "
                 f"Input '{input_name}' has concept '{input_requirement.concept.concept_string}'. "
                 "Image generation requires a text prompt. Use native.Text or a concept that refines it."
             )
             raise PipeValidationError(
+                message=msg,
                 error_type=PipeValidationErrorType.IMG_GEN_INPUT_NOT_TEXT_COMPATIBLE,
                 pipe_code=self.code,
                 variable_names=[input_name],
                 required_concept_codes=["native.Text"],
                 provided_concept_code=input_requirement.concept.concept_string,
-                explanation=explanation,
             )
 
     @override
@@ -149,16 +149,17 @@ class PipeImgGen(PipeOperator[PipeImgGenOutput]):
             wanted_concept=get_native_concept(native_concept=NativeConceptCode.IMAGE),
             strict=True,
         ):
+            msg = (
+                f"The output of a ImgGen pipe must be compatible with the Image concept. "
+                f"In the pipe '{self.code}' the output is '{self.output.concept_string}'"
+            )
             raise PipeValidationError(
+                message=msg,
                 error_type=PipeValidationErrorType.INADEQUATE_OUTPUT_CONCEPT,
                 domain=self.domain,
                 pipe_code=self.code,
                 provided_concept_code=self.output.concept_string,
                 required_concept_codes=[NativeConceptCode.IMAGE.concept_string],
-                explanation=(
-                    f"The output of a ImgGen pipe must be compatible with the Image concept. "
-                    f"In the pipe '{self.code}' the output is '{self.output.concept_string}'"
-                ),
             )
 
     @override
